+++ src/compas/scene/scene.py
@@ -3,9 +3,8 @@ from compas.datastructures import Tree
 from compas.datastructures import TreeNode
 
 from .context import clear
-from .context import predraw
-from .context import postdraw
-from .context import redraw
+from .context import before_draw
+from .context import after_draw
 from .context import detect_current_context
 from .sceneobject import SceneObject
 
@@ -201,7 +200,7 @@ class Scene(Data):
     >>> scene = Scene()
     >>> box = Box.from_width_height_depth(1, 1, 1)
     >>> scene.add(box)
-    >>> scene.redraw()
+    >>> scene.draw()
 
     """
 
@@ -292,7 +291,7 @@ class Scene(Data):
     def draw(self):
         """Draw the scene."""
 
-        predraw()
+        before_draw()
 
         if not self.context:
             raise ValueError("No context detected.")
@@ -303,14 +302,9 @@ class Scene(Data):
         for sceneobject in self.objects:
             drawn_objects += sceneobject.draw()
 
-        if drawn_objects:
-            redraw()
-
-        postdraw()
+        after_draw(drawn_objects)
 
         return drawn_objects
-
-    redraw = draw
 
     def print_hierarchy(self):
         """Print the hierarchy of the scene."""