+++ CHANGELOG.md
@@ -12,8 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `compas_rhino.objects`.
 * Added `compas_rhino.layers`.
 * Added `compas_rhino.install_with_pip`.
-* Added `predraw` pluggable to `compas.scene.Scene.draw`.
-* Added `postdraw` pluggable to `compas.scene.Scene.draw`.
+* Added `before_draw` pluggable to `compas.scene.Scene.draw`.
+* Added `after_draw` pluggable to `compas.scene.Scene.draw`.
 
 ### Changed